Improve code reuse in Reactivity, add tests
