+++ packages/reactivity/package.json
@@ -14,7 +14,12 @@
     "scripts": {
         "build": "npm run build:esm && npm run build:browser",
         "build:browser": "rollup -c rollup.browser.config.mjs",
-        "build:esm": "rollup -c rollup.esm.config.mjs"
+        "build:esm": "rollup -c rollup.esm.config.mjs",
+        "watch": "rollup -c rollup.esm.config.js -w",
+        "lint": "eslint -c ../../.eslintrc.json --ext .js src tests",
+        "test": "jest -c ../../jest.config.js ./tests/**/*.test.js",
+        "prepare": "husky install",
+        "prepublishOnly": "npm run build"
     },
     "author": "Ben Thomson <git@alfreido.com>",
     "homepage": "https://wintercms.com",
