+++ packages/reactivity/tests/fixtures/TestSnowboardTemplate.js
@@ -1,0 +1,14 @@
+import ReactiveSingleton from '../../src/abstracts/ReactiveSingleton';
+
+export default class TestSnowboardTemplate extends ReactiveSingleton {
+    construct() {
+    }
+
+    template() {
+        return `
+            <div>
+                <strong>{{ snowboard.plugins }}</strong>
+            </div>
+        `;
+    }
+}
